Beta Demo

See merge request mentoru/mentoru!14

+++ MentorU/MentorU/MentorU/ViewModels/NewItemViewModel.cs
@@ -6,6 +6,10 @@ using MentorU.Services.DatabaseServices;
 using System;
 using System.IO;
 using Xamarin.Forms;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using Plugin.Media;
+using Plugin.Media.Abstractions;
 
 namespace MentorU.ViewModels
 {
@@ -26,6 +30,16 @@ namespace MentorU.ViewModels
             CancelCommand = new Command(OnCancel);
             this.PropertyChanged +=
                 (_, __) => SaveCommand.ChangeCanExecute();
+            AllDepartments = new List<string>(DatabaseService.Instance.ClassList.classList);
+            Department = AllDepartments[0];
+
+            AllConditions = new ObservableCollection<string>()
+            {
+                "New",
+                "Like New",
+                "Good",
+                "Decent"
+            };
         }
 
         public ImageSource ItemFirstImage
@@ -41,14 +55,36 @@ namespace MentorU.ViewModels
         private async void AddPicture()
         {
 
-            Stream itemImageStream = await DependencyService.Get<IPhotoPickerService>().GetImageStreamAsync();
-            if (itemImageStream != null)
-            {
-                string fileName = "temp-market-image";
-                itemImageFilePath = DependencyService.Get<IFileService>().SavePicture(fileName, itemImageStream);
+            //Stream itemImageStream = await DependencyService.Get<IPhotoPickerService>().GetImageStreamAsync();
+            //if (itemImageStream != null)
+            //{
+            //    string fileName =
+            //    itemImageFilePath = DependencyService.Get<IFileService>().SavePicture(fileName, itemImageStream);
 
-                ItemFirstImage = itemImageFilePath;
+            //    ItemFirstImage = itemImageFilePath;
+            //}
+
+
+
+            await CrossMedia.Current.Initialize();
+            if (!CrossMedia.Current.IsPickPhotoSupported)
+            {
+                await AppShell.Current.DisplayAlert("Not supported", "Your device does not currently support this functionality", "Ok");
+                return;
             }
+
+            var mediaOption = new PickMediaOptions()
+            {
+                PhotoSize = PhotoSize.Medium
+            };
+
+            string fileName = "temp-market-image";
+
+            var selectedImageFile = await CrossMedia.Current.PickPhotoAsync(mediaOption);
+
+            itemImageFilePath = DependencyService.Get<IFileService>().SavePicture(fileName, selectedImageFile.GetStream());
+
+            ItemFirstImage = itemImageFilePath;
 
         }
 
@@ -76,6 +112,44 @@ namespace MentorU.ViewModels
             set => SetProperty(ref description, value);
         }
 
+
+        public List<string> AllDepartments { get; set; }
+
+        private string _department;
+        public string Department
+        {
+            get => _department;
+            set
+            {
+                _department = value;
+                OnPropertyChanged();
+            }
+        }
+
+        private string _classNumber;
+        public string ClassNumber
+        {
+            get => _classNumber;
+            set
+            {
+                _classNumber = value;
+                OnPropertyChanged();
+            }
+        }
+
+        public ObservableCollection<string> AllConditions { get; set; }
+
+        private string _conditions;
+        public string condition
+        {
+            get => _conditions;
+            set
+            {
+                _conditions = value;
+                OnPropertyChanged();
+            }
+        }
+
         public Command SaveCommand { get; }
         public Command CancelCommand { get; }
 
@@ -92,24 +166,27 @@ namespace MentorU.ViewModels
                 Text = Text,
                 Description = Description,
                 Price = ItemPrice,
-                Owner = App.loggedUser.id
+                Owner = App.loggedUser.id,
+                ClassUsed = Department + " " + ClassNumber,
+                Condition = condition
             };
 
             await DatabaseService.Instance.client.GetTable<Items>().InsertAsync(newItem);
 
             string containeritemid = newItem.id;
 
+            if(itemImageFilePath != null)
+            {
+                BlobContainerClient containerClient = BlobService.Instance.BlobServiceClient.GetBlobContainerClient(containeritemid);
+                await containerClient.CreateIfNotExistsAsync();
 
-            BlobContainerClient containerClient = BlobService.Instance.BlobServiceClient.GetBlobContainerClient(containeritemid);
-            await containerClient.CreateIfNotExistsAsync();
+                string fileName = "Image0";
 
-            string fileName = "Image0";
+                //deletes the blob file if it exists and uploads an image
+                await BlobService.Instance.TryUploadImage(containerClient, fileName, itemImageFilePath);
 
-            //deletes the blob file if it exists and uploads an image
-            await BlobService.Instance.TryUploadImage(containerClient, fileName, itemImageFilePath);
-
-
-            File.Delete(itemImageFilePath);
+                File.Delete(itemImageFilePath);
+            }
 
             await Application.Current.MainPage.DisplayAlert("Success", "Item Added", "Ok");
